backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
ultralytics==8.3.0
torch==2.5.1
torchvision==0.20.1
opencv-python-headless==4.10.0.84
Pillow==10.4.0
python-dotenv==1.0.1
pymongo==4.8.0
boto3==1.34.0
tensorflow==2.17.0
shapely==2.0.6